requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=3.13
Pillow>=9.0
pytesseract>=0.3.10
numpy>=1.23
